techDebt(registrationRefactor): Move responseErrors and RegistrationException into separate file
