+++ server/routes/registration/registerAccount.js
@@ -10,87 +10,9 @@ const EstablishmentSaveException =
   require('../../models/classes/establishment/establishmentExceptions').EstablishmentSaveException;
 const UserModel = require('../../models/classes/user').User;
 const UserSaveException = require('../../models/classes/user/userExceptions').UserSaveException;
+const { responseErrors, RegistrationException } = require('./responseErrors');
 
 const OTHER_MAX_LENGTH = 120;
-
-const responseErrors = {
-  default: {
-    errCode: -1,
-    errMessage: 'Registration Error',
-  },
-  eastablishment: {
-    errCode: -2,
-    errMessage: 'Registration: Failed to create Establishment',
-  },
-  user: {
-    errCode: -3,
-    errMessage: 'Registration: Failed to create User',
-  },
-  login: {
-    errCode: -4,
-    errMessage: 'Registration: Failed to create Login',
-  },
-  duplicateNonCQC: {
-    errCode: -100,
-    errMessage: 'Duplicate non-CQC Establishment',
-    db_constraint: 'Establishment_unique_registration',
-  },
-  duplicateCQC: {
-    errCode: -150,
-    errMessage: 'Duplicate CQC Establishment',
-    db_constraint: 'Establishment_unique_registration_with_locationid',
-  },
-  duplicateEstablishment: {
-    errCode: -190,
-    errMessage: 'Duplicate Establishment',
-  },
-  duplicateUsername: {
-    errCode: -200,
-    errMessage: 'Duplicate Username',
-    db_constraint: 'uc_Login_Username',
-  },
-  unexpectedMainService: {
-    errCode: -300,
-    errMessage: 'Unexpected main service',
-  },
-  unknownLocation: {
-    errCode: -400,
-    errMessage: 'Unknown location',
-    db_constraint: 'uc_Login_Username',
-  },
-  unknownNMDSsequence: {
-    errCode: -500,
-    errMessage: 'Unknown NMDS Sequence',
-  },
-  unknownNMDSLetter: {
-    errCode: -600,
-    errMessage: 'Unknown NMDS Letter/CSSR Region',
-  },
-  invalidEstablishment: {
-    errCode: -700,
-    errMessage: 'Establishment data is invalid',
-  },
-  invalidUser: {
-    errCode: -800,
-    errMessage: 'User data is invalid',
-  },
-  invalidUsername: {
-    errCode: -210,
-    errMessage: 'Invalid Username',
-  },
-};
-
-class RegistrationException {
-  constructor(originalError, errCode, errMessage) {
-    this.err = originalError;
-    this.errCode = errCode;
-    this.errMessage = errMessage;
-  }
-
-  toString() {
-    return `${this.errCode}: ${this.errMessage}`;
-  }
-}
 
 exports.registerAccount = async (req, res) => {
   //basic validation
